[refactor/clean-js-debounce] move debounce to utils file
