+++ openlibrary/plugins/openlibrary/js/nonjquery_utils.js
@@ -1,0 +1,22 @@
+/**
+ * These methods are separate from utils.js because that file assumes a globel jQuery object.
+ */
+
+export function debounce(func, threshold, execAsap) {
+    var timeout;
+    return function debounced () {
+        var obj = this, args = arguments;
+        function delayed () {
+            if (!execAsap)
+                func.apply(obj, args);
+            timeout = null;
+        }
+
+        if (timeout) {
+            clearTimeout(timeout);
+        } else if (execAsap) {
+            func.apply(obj, args);
+        }
+        timeout = setTimeout(delayed, threshold || 100);
+    };
+};
